+++ gradle.properties
@@ -1,6 +1,6 @@
 # Gradle properties
-pluginName = CivBoats
-group = com.aleksey.civboats
-version = 1.0.3
+pluginName = CivVehicles
+group = com.aleksey.civvehicles
+version = 1.1.0
 description = Boats with inventory
 api_version=1.18
